Extract an ErlParser interface
