+++ test/gazelle/fetch_test.go
@@ -3,8 +3,8 @@ package fetch_test
 import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
-	"testing"
 	"github.com/rabbitmq/rules_erlang/gazelle/fetch"
+	"testing"
 )
 
 func TestUpdate(t *testing.T) {
@@ -13,7 +13,7 @@ func TestUpdate(t *testing.T) {
 }
 
 var _ = Describe("ParseGithubImportArg", func() {
-	It("should parse", func() {
+	It("should parse the full specification", func() {
 		imp := "inet_tcp_proxy_dist=github.com/rabbitmq/inet_tcp_proxy@master"
 		name, owner, repo, ref, err := fetch.ParseGithubImportArg(imp)
